ai/pyproject.toml
[project]
name = "ol-rf-mesh-ai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scikit-learn",
    "joblib",
    "tensorflow-cpu>=2.14.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
